refactor: extract color/theme utilities from cli.py to hermes_cli/cli_theme.py
